pyproject.toml
[project]
name = "codereviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "dotenv>=0.9.9",
    "fastapi>=0.129.0",
    "google-genai>=1.64.0",
    "httpx>=0.28.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
